Make ema more precise

+++ uvw_iot/util/Filter.h
@@ -4,6 +4,6 @@
 
 namespace uvw_iot::util {
 
-void ema(int& prev, int curr, std::chrono::milliseconds deltaT, std::chrono::milliseconds tau);
+void ema(double& prev, int curr, std::chrono::milliseconds deltaT, std::chrono::milliseconds tau);
 
 } // namespace uvw_iot::util
